Fix type issue for mesh field and 2D field reader.
This fixes an issue with the type for kiteAreasOnVertex and the 2D
field reading helper.

+++ mpasoflecsi/specialization/mesh.hh
@@ -129,7 +129,7 @@ inline const flecsi::field<double>::definition<mesh, mesh::vertices> fVertex;
 inline const flecsi::field<double>::definition<mesh, mesh::edges>    fEdge;
 inline const flecsi::field<double>::definition<mesh, mesh::cells>    bottomDepth;
 
-inline const flecsi::field<vdtensor<int>>::definition<mesh, mesh::vertices> kiteAreasOnVertex;
+inline const flecsi::field<vdtensor<double>>::definition<mesh, mesh::vertices> kiteAreasOnVertex;
 inline const flecsi::field<me2tensor<double>>::definition<mesh, mesh::edges> weightsOnEdge;
 
 }
